Fix compiling

+++ Source/Core/Core.cpp
@@ -254,7 +254,7 @@ Context* CreateContext(const String& name, const Vector2i& dimensions, RenderInt
 	if (dimensions.x > 0 && dimensions.y > 0)
 	{
 		// install an orthographic projection, by default
-		Matrix4f P = Matrix4f::ProjectOrtho(0, dimensions.x, dimensions.y, 0, -1, 1);
+		Matrix4f P = Matrix4f::ProjectOrtho(0, (float)dimensions.x, (float)dimensions.y, 0, -1, 1);
 		new_context->ProcessProjectionChange(P);
 		// install an identity view, by default
 		new_context->ProcessViewChange(Matrix4f::Identity());
